requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.5
django-appconf==1.0.4
django-cors-headers==3.6.0
django-etc==1.3.0
django-hitcount==1.3.3
django-imagekit==4.0.2
django-tinymce4-lite==1.8.0
gunicorn==20.0.4
idna==2.10
jsmin==2.2.2
pilkit==2.0
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
cloudinary==1.24.0
django-cloudinary-storage==0.3.0
